+++ pytest/test_io_meshes.py
@@ -55,33 +55,21 @@ def generate_pygmsh_sphere():
 
 @pytest.mark.parametrize("generate_pygmsh",
                          [generate_pygmsh_cylinder, generate_pygmsh_sphere, generate_pygmsh_wavebot])
-def test_from_meshio_pygmsh(generate_pygmsh, tmp_path):
+def test_from_meshio_pygmsh(generate_pygmsh):
     pytest.importorskip("pygmsh", reason="PyGMSH not installed, test skipped")
     mesh, vol_exp = generate_pygmsh()
-    fb = cpt.FloatingBody(mesh=mesh)
-    fb.mesh.heal_mesh()
-    fb.keep_immersed_part()
+    mesh = cpt.load_mesh(mesh)
+    assert mesh.immersed_part().volume == pytest.approx(vol_exp, rel=1e-1)
 
-    vol = fb.mesh.volume
-    assert pytest.approx(vol_exp, rel=1e-1) == vol
-
-    fb.add_translation_dof((0,0,1))
-
+    fb = cpt.FloatingBody(mesh)
+    fb.add_translation_dof(name="Heave")
     test_matrix = xr.Dataset(coords={
-        'rho': [1e3],
-        'water_depth': [np.inf],
         'omega': [np.pi],
-        'wave_direction': 0,
+        'wave_direction': [0],
         'radiating_dof': list(fb.dofs.keys()),
         })
-
     solver = cpt.BEMSolver()
-    solver.fill_dataset(test_matrix,
-                        bodies=[fb],
-                        hydrostatics=True,
-                        mesh=True,
-                        wavelength=True,
-                        wavenumber=True)
+    solver.fill_dataset(test_matrix, bodies=fb)
 
 #################################################################